Going through all of our subsystems
We updated all of the subsystems to reflect how we want to use them.
We also went through most of our commands, and command groups, to get them ready as well. We will finish chekcing the commands by the end of the week, and add in finishing touches as we go.